+++ dhcp/Dockerfile
@@ -1,12 +1,28 @@
 # FROM balenalib/raspberrypi3-alpine-node:latest
 FROM node:latest
 
-WORKDIR /usr/src/app
+# Install yarn
+RUN npm i --force -g yarn
 
-COPY package*.json ./
-RUN npm i
+# Copy package.json
+WORKDIR /app/dhcp
+COPY ./dhcp/package.json .
+RUN yarn install
 
-COPY . .
+# Copy files
+COPY dhcp .
+
+# Install utils
+WORKDIR /app/utils
+COPY utils .
+RUN yarn install
+WORKDIR /app/dhcp
+
+# Build
+RUN yarn build
+
+RUN ls -l built/utils
+RUN ls -l built/utils/msgMgr
 
 EXPOSE 67
 CMD ["node", "index.js"]
