+++ dhcp/package.json
@@ -25,6 +25,7 @@
   },
   "devDependencies": {
     "@types/node": "^14.0.6",
+    "typescript": "^3.8.3",
     "mocha": "^7.2.0",
     "should": "^13.2.3"
   }
