--- msg_broker/src/msgMgr/msgMgr.d.ts
@@ -1,6 +1,0 @@
-export interface MessageManager {
-    connect(): void;
-    readMsg(queue?: string): Promise<string>;
-}
-
-declare module "MsgMgr";
