+++ pcap/package.json
@@ -4,19 +4,19 @@
   "description": "Reads all packets to list services used on network",
   "main": "index.js",
   "scripts": {
-    "start": "node ./built/index.js",
-    "build": "tsc && ncc build built --minify -o prod",
+    "start": "node ./built/pcap/src/index.js",
+    "build": "tsc",
     "beautify": "prettier --config ../.prettier.json ./**/*.js --write",
     "test": "mocha"
   },
   "author": "Vivi",
   "license": "MIT",
   "dependencies": {
-    "@types/amqplib": "^0.5.13",
     "amqplib": "^0.5.5",
     "pcap": "^3.1.0"
   },
   "devDependencies": {
+    "@types/amqplib": "^0.5.13",
     "@types/node": "^13.13.12",
     "mocha": "^7.2.0",
     "should": "^13.2.3",
