--- pcap/src/msgMgr/msgMgr.d.ts
@@ -1,7 +1,0 @@
-export interface MessageManager {
-    connect(): void;
-    sendMsg(message: string, queue?: string): void;
-    sendErr(error: string): void;
-}
-
-declare module "MsgMgr";
