+++ utils/msgMgr/msgMgr.d.ts
@@ -1,5 +1,6 @@
 export interface MessageManager {
     connect(): void;
+    readMsg(queue?: string): Promise<string>;
     sendMsg(message: string, queue?: string): void;
     sendErr(error: string): void;
 }
